Code simplification, removal of 'exact'
I simplified some code in areas that aren't performance bottlenecks. Also, I implemented the 'exact' linear sum assignment solver, and it turns out that I had made a miscalculation and the reason it appears to halve the path lengths was because I neglected to count the movements happening inside the target square. In reality, the difference was negligible, and so I'll stick with the greedy algorithm currently in use.