More simplifications
I have simplified the algorithm for finding the nearest free atom to a hole, it now uses scipy's `cdist` function to calculate all pairs of distances between the two sets, then sorts them and iterates through them. There are O(N) free atoms, and O(N^(1/2)) holes (as the length of the perimeter is O(N^(1/2)), so O(N^(3/2)) distances. The sorting is then O(NlogN), so the overall calculation is O(N^(3/2)logN). Despite that, it is faster than the old version in practice, and the sorting step is nowhere near the runtime bottleneck.