Change Python 3.11 specific syntax, fix 'pad = False'
I accidentally used some syntax specific to Python 3.11, so I've changed it to avoid doing that. I also fixed the 'pad = False' option to work as intended. This option essentially forces all clusters to be connected internally, rather than counting a connection to the perimeter. This requires more moves (0.57 for small cases, but limits to the same value around 0.54), but it reduces path lengths by around 20% on average and 50% in the worst case. It seems more effective at larger sizes, so N > 10,000.