Optimised `push_to_bottom`, better var names
`push_to_bottom` now caches the last path used, so when multiples atoms are pushed down the same path it doesn't have to recalculate the whole thing. This speeds up this function in particular by a factor of around 4, so now the total runtime is split fairly evenly between `push_to_bottom`, `move_to_hole` and the path length calculation at the end.